+++ src/event-listeners/KeyUpEventListener.ts
@@ -7,7 +7,12 @@ class KeyUpEventListener extends TargetedEventListener {
   userActionType = UserActionType.KeyUp
 
   listener(event: KeyboardEvent) {
-    const userAction = createTargetedUserAction(event, this.userActionType, this.options)
+    const userAction = createTargetedUserAction(
+      event,
+      this.userActionType,
+      this.userActionHandler.getAnonymizationSettings(),
+      this.options,
+    )
     if (userAction === null) return
     if (isKeyAllowedByKeyListeners(event.code)) {
       return this.userActionHandler.handle(userAction)