+++ src/user-action/UserActionHandler.ts
@@ -1,4 +1,4 @@
-import { SessionUserAction, UserAction } from '../types'
+import { AnonymizationSettings, NO_ANONYMIZATION_SETTINGS, SessionUserAction, UserAction } from '../types'
 import ISessionIdHandler from '../session-id-handler/ISessionIdHandler'
 import IUserActionHandler from './IUserActionHandler'
 import { IGravityClient } from '../gravity-client/IGravityClient'
@@ -7,12 +7,21 @@ import ITimeoutHandler from '../timeout-handler/ITimeoutHandler'
 export default class UserActionHandler implements IUserActionHandler {
   private active = true
   private readonly listeners: Array<Function> = []
+  private anonymizationSettings: AnonymizationSettings = NO_ANONYMIZATION_SETTINGS
 
   constructor(
     private readonly sessionIdHandler: ISessionIdHandler,
     private readonly timeoutHandler: ITimeoutHandler,
     private readonly gravityClient: IGravityClient,
   ) {}
+
+  getAnonymizationSettings(): AnonymizationSettings {
+    return this.anonymizationSettings
+  }
+
+  setAnonymizationSettings(anonymizationSettings: AnonymizationSettings) {
+    this.anonymizationSettings = anonymizationSettings
+  }
 
   async handle(action: UserAction): Promise<void> {
     if (!this.active) return